requirements.txt
python-dotenv==1.0.0
azure-ai-documentintelligence==1.0.0b3